Fix resetting of reverse order checkbox, order is calculated now in service
