+++ client/src/services/search.tsx
@@ -2,13 +2,33 @@ import axios from 'axios';
 import getAuthHeader from './auth';
 import { handleAxiosError } from '../utils/errors';
 import { apiBaseUrl } from '../constants';
-import { SortAndFilter } from '../types';
+import { SortAndFilter, SortingCriteria } from '../types';
+
+const getOrder = (value: string, reversed: boolean): SortingCriteria => {
+	switch (value) {
+		case 'distance':
+			if (reversed) return { sort: 'distance', order: 'desc' };
+			return { sort: 'distance', order: 'asc' };
+		case 'age':
+			if (reversed) return { sort: 'age', order: 'desc' };
+			return { sort: 'age', order: 'asc' };
+		case 'rating':
+			if (reversed) return { sort: 'rating', order: 'asc' };
+			return { sort: 'rating', order: 'desc' };
+		case 'tags':
+			if (reversed) return { sort: 'tags', order: 'asc' };
+			return { sort: 'tags', order: 'desc' };
+		default:
+			return { sort: 'distance', order: 'asc' };
+	}
+};
 
 export const getMatchSuggestions = async (sortAndFilter: SortAndFilter): Promise<any> => {
-	const { distance, age, rating, tags } = sortAndFilter.filter;
 
+	const { sort, isReversedOrder } = sortAndFilter.sort;
+	const { distance, age, rating, tags } = sortAndFilter.filter;
 	const sortAndFilterCriterias = {
-		sort: sortAndFilter.sort,
+		sort: getOrder(sort, isReversedOrder),
 		filter: [
 			{ filter: 'distance', min: distance.min, max: distance.max },
 			{ filter: 'age', min: age.min, max: age.max },
